refactor : update profile
